requirements.txt
google-genai
anthropic
pandas
pyarrow
pillow
python-dotenv
